Fix #623 by repackaging Responder in Bolt
