+++ bolt/src/main/java/com/slack/api/bolt/context/builtin/ActionContext.java
@@ -2,7 +2,7 @@ package com.slack.api.bolt.context.builtin;
 
 import com.slack.api.bolt.context.ActionRespondUtility;
 import com.slack.api.bolt.context.Context;
-import com.slack.api.bolt.response.Responder;
+import com.slack.api.bolt.util.Responder;
 import lombok.*;
 
 /**
