+++ bolt/src/main/java/com/slack/api/bolt/context/builtin/SlashCommandContext.java
@@ -4,7 +4,7 @@ import com.slack.api.app_backend.slash_commands.response.SlashCommandResponse;
 import com.slack.api.bolt.context.Context;
 import com.slack.api.bolt.context.RespondUtility;
 import com.slack.api.bolt.context.SayUtility;
-import com.slack.api.bolt.response.Responder;
+import com.slack.api.bolt.util.Responder;
 import com.slack.api.bolt.response.Response;
 import com.slack.api.bolt.util.BuilderConfigurator;
 import com.slack.api.model.block.LayoutBlock;
